Fix username and add dimension to result
This fixes an issue where if a player is on a team with a colored name, no name will be shown in chat. Instead, we display the raw username without any formatting or coloring. Not sure if there is a better way but this worked for me. Additionally, this PR adds the current dimension of the player to the result which is also helpful.